Restructured integration test suite
